+++ iappslx/presentation/app.js
@@ -71,6 +71,7 @@ let UI;
 
 const appState = {
     debugOutput: '',
+    foundAS3: true,
     data: {},
     modal: {
         message: '',
@@ -283,7 +284,14 @@ const newEditor = (tmplid, view) => {
                 dispOutput(JSON.stringify(tmpl.getCombinedParameters(editor.getValue()), null, 2));
             };
             document.getElementById('view-render-btn').onclick = () => {
-                dispOutput(tmpl.render(editor.getValue()));
+                const rendered = tmpl.render(editor.getValue());
+                const msg = [
+                    'WARNING: The below declaration is only for inspection and debug purposes. Submitting the ',
+                    'below ouput to AS3 directly can result in loss of tenants\nand applications. Please only ',
+                    'submit this declaration through FAST.\n\n',
+                    rendered
+                ].join('');
+                dispOutput(msg);
             };
             document.getElementById('btn-form-submit').onclick = () => {
                 const parameters = editor.getValue();
@@ -330,6 +338,14 @@ const newEditor = (tmplid, view) => {
             }
         });
 };
+
+// Check that AS3 is available
+safeFetch('/mgmt/shared/appsvcs/info')
+    .catch((e) => {
+        appState.foundAS3 = false;
+        appState.busy = false;
+        console.log(`Error reaching AS3: ${e.message}`);
+    });
 
 // Setup basic routing
 const routes = {};
@@ -547,6 +563,13 @@ route('templates', 'templates', () => {
                     console.error('enabled === false && updateAvailable is illegal. Critical Error');
                 }
             });
+
+            appState.data.errors = disabledTemplateSets.reduce((acc, curr) => {
+                if (curr.error) {
+                    acc.push(curr.error);
+                }
+                return acc;
+            }, []);
         })
         .then(() => dispOutput(''));
 
@@ -702,7 +725,8 @@ route('templates', 'templates', () => {
         filters,
         currentFilter,
         sets: [],
-        apps: {}
+        apps: {},
+        errors: []
     };
 
     return renderTemplates();