+++ test/unit/worker.js
@@ -10,6 +10,7 @@ const url = require('url');
 
 process.AFL_TW_ROOT = path.join(process.cwd(), './templates');
 process.AFL_TW_TS = path.join(process.cwd(), './templates');
+delete process.env.FAST_BIGIP_HOST; // Never try targeting a remote BIG-IP
 
 const fs = require('fs');
 const assert = require('assert').strict;
@@ -139,9 +140,9 @@ function createWorker() {
 }
 
 describe('template worker tests', function () {
-    const host = 'http://localhost:8105';
-    const as3ep = '/shared/appsvcs/declare';
-    const as3TaskEp = '/shared/appsvcs/task';
+    const host = 'http://localhost:8100';
+    const as3ep = '/mgmt/shared/appsvcs/declare';
+    const as3TaskEp = '/mgmt/shared/appsvcs/task';
     const as3stub = {
         class: 'ADC',
         schemaVersion: '3.0.0'
@@ -384,8 +385,8 @@ describe('template worker tests', function () {
                 [AS3DriverConstantsKey]: appData
             }
         };
-        nock('http://localhost:8100')
-            .get('/mgmt/shared/appsvcs/declare')
+        nock(host)
+            .get(as3ep)
             .reply(200, Object.assign({}, as3stub, {
                 tenant: {
                     class: 'Tenant',
@@ -532,7 +533,7 @@ describe('template worker tests', function () {
         return worker.onPost(op)
             .then(() => {
                 assert.equal(op.status, 404);
-                assert.match(op.body.message, /No templates found/);
+                assert.match(op.body.message, /Could not find template/);
             });
     });
     it('post_apps_bad_tmplid_leading_slash', function () {