+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # 1.6.0
 ## Added
+
 ## Fixed
+
 ## Changed
+* GUI: Display an error if AS3 is not available
+* GUI: Report template errors in the Template tab
+* bigip-fast-templates: Add a default pool member
 
 # 1.5.0
 ## Added